broadcom/vc4: Add support for extended CL submission
Recent versions of the VC4 driver support CL extensions, which allows one
to pass extra attributes to the CL submission.
In order to make this feature backward compatible with the existing
SUBMIT_CL ioctl, we had to add a new flag and re-use the optional
bin_cl fields. Binner CL are now passed as an extra chunk.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>