panfrost: Fix ->set_vertex_buffers() for partial vertex bufs updates
The caller of ->set_vertex_buffers() might want to update only a subset
of the currently active vertex buffers, hence the start_slot and
num_buffers arguments.
The panfrost driver was assuming that the whole set of vertex buffers
were to be freed/updated each time this hook is called, which leads to
issues when the caller want to update only some of those buffers (this
is what the gallium-cso layer does when it saves/restores the vertex
buf context).

Fix the logic to allow partial vertex bufs updates.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
---
Hi,

I faced this issue while trying to make gallium-hud work with the
panfrost driver. I'm still struggling to fix at least 3 remaining
issues:

- the Y coordinates are inverted, and the HUD is displayed on the
  bottom-left corner instead of top-left corner, plus it's mirrored
- transparency/alpha-blending does not work correctly (I get black
  squares instead of semi transparent grey ones)
- the fonts shader (or something related to it) does not seem to work
  (no legend on the graph)

If anyone has an idea where this could come from, please let me know.

Thanks,

Boris